@@ -1,4 +1,4 @@
-// swift-tools-version:5.7
+// swift-tools-version: 6.0
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -6,8 +6,8 @@ import PackageDescription
 let package = Package(
     name: "APIKeyReader",
     platforms: [
-        .iOS(.v16),
-        .macOS(.v13),
+        .iOS(.v18),
+        .macOS(.v15),
     ],
     products: [
         // Products define the executables and libraries a package produces, and make them visible to other packages.
@@ -17,14 +17,14 @@ let package = Package(
         ),
     ],
     dependencies: [
-        .package(url: "https://github.com/kraigspear/Spearfoundation.git", .exact("2.0.1"))
+       
     ],
     targets: [
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
         // Targets can depend on other targets in this package, and on products in packages this package depends on.
         .target(
             name: "APIKeyReader",
-            dependencies: [.product(name: "SpearFoundation", package: "Spearfoundation")]
+            dependencies: []
         ),
         .testTarget(
             name: "APIKeyReaderTests",
